move do_complete to SpyderPdb
