+++ spyder_kernels/console/kernel.py
@@ -19,7 +19,6 @@ from ipykernel.ipkernel import IPythonKernel
 
 # Local imports
 from spyder_kernels.comms.frontendcomm import FrontendComm
-from spyder_kernels.py3compat import isidentifier, PY2
 
 
 # Excluded variables from the Variable Explorer (i.e. they are not
@@ -255,50 +254,8 @@ class SpyderKernel(IPythonKernel):
 
         Public method of ipykernel overwritten for debugging.
         """
-        if cursor_pos is None:
-            cursor_pos = len(code)
         if self.is_debugging():
-            # Get text to complete
-            text = code[:cursor_pos].split(' ')[-1]
-            # Choose pdb function to complete, based on cmd.py
-            origline = code
-            line = origline.lstrip()
-            if not line:
-                return
-            stripped = len(origline) - len(line)
-            begidx = cursor_pos - len(text) - stripped
-            endidx = cursor_pos - stripped
-            if begidx > 0:
-                cmd, args, _ = self._pdb_obj.parseline(line)
-                if cmd == '':
-                    compfunc = self._pdb_obj.completedefault
-                else:
-                    try:
-                        compfunc = getattr(self._pdb_obj, 'complete_' + cmd)
-                    except AttributeError:
-                        compfunc = self._pdb_obj.completedefault
-            elif line[0] != '!':
-                compfunc = self._pdb_obj.completenames
-            else:
-                compfunc = self._pdb_obj.completedefault
-
-            def is_name_or_composed(text):
-                if not text or text[0] == '.':
-                    return False
-                # We want to keep value.subvalue
-                return isidentifier(text.replace('.', ''))
-
-            while text and not is_name_or_composed(text):
-                text = text[1:]
-                begidx += 1
-
-            matches = compfunc(text, line, begidx, endidx)
-
-            return {'matches': matches,
-                    'cursor_end': cursor_pos,
-                    'cursor_start': cursor_pos - len(text),
-                    'metadata': {},
-                    'status': 'ok'}
+            return self._pdb_obj.do_complete(code, cursor_pos)
         return super(SpyderKernel, self).do_complete(code, cursor_pos)
 
     def publish_pdb_state(self):
